requirements.txt
folium
streamlit_folium
scikit-learn